Commit fix encode and decode JSON

+++ app/src/main/java/com/br/highbee/models/ProductsHome.kt
@@ -2,4 +2,11 @@ package com.br.highbee.models
 
 import java.math.BigDecimal
 
-data class ProductsHome(val id: Int, val name: String, val img: String, val price: BigDecimal)
+
+
+data class ProductsHome(
+    val id: Int,
+    val name: String,
+    val img: String,
+    val price: Double
+)
